+++ lib/src/components/components.dart
@@ -1,5 +1,5 @@
 export 'bottom_bar.dart';
 export 'defines.dart';
 export 'dialogs.dart';
-export 'member/member_button.dart';
+export 'member/button.dart';
 export 'top_bar.dart';