+++ lib/src/core/plugins.dart
@@ -228,7 +228,9 @@ class ZegoPrebuiltPlugins {
       if (ZegoPluginAdapter().getPlugin(ZegoUIKitPluginType.signaling) !=
           null) {
         await ZegoUIKit().getSignalingPlugin().leaveRoom();
-        await ZegoUIKit().getSignalingPlugin().logout();
+
+        /// not need logout
+        // await ZegoUIKit().getSignalingPlugin().logout();
         await ZegoUIKit().getSignalingPlugin().uninit();
       }
     }